Capturing "find parking by ADDRESS" scenario with search address functionality from Azure Maps
